Renamed unused Gallery Fragment to Notifications Fragment and added icon on bottom of screen. This fragment can now be built upon for Notifications/Activity History
Deleted more outdated files: slideshowFragment and fragment_slideshow. (unused fragment)
Deleted: activity_main_drawer.xml, activity_nav_drawer.xml, nav_drawer.xml (all files were for the old design with the icons on the side)
