+++ app/src/main/java/com/example/warriorsocial/ui/notifications/NotificationsViewModel.java
@@ -1,16 +1,16 @@
-package com.example.warriorsocial.ui.gallery;
+package com.example.warriorsocial.ui.notifications;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class GalleryViewModel extends ViewModel {
+public class NotificationsViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public GalleryViewModel() {
+    public NotificationsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is gallery fragment");
+        mText.setValue("This is the notifications fragment");
     }
 
     public LiveData<String> getText() {
