+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -17,6 +17,11 @@
         android:title="Discover" />
 
     <item
+        android:id="@+id/navigation_notifications"
+        android:icon="@drawable/ic_notifications_black_24dp"
+        android:title="Notifications" />
+
+    <item
         android:id="@+id/navigation_settings"
         android:icon="@drawable/ic_baseline_settings_24"
         android:title="Settings" />
