--- app/src/main/res/menu/nav_drawer.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <!-- This page is for the drop down from the 3 dots -->
-
-    <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        android:title="@string/action_settings"
-        app:showAsAction="never" />
-    <item
-        android:id="@+id/action_anotherthing"
-        android:orderInCategory="100"
-        android:title="Another thing"
-        app:showAsAction="never" />
-</menu>
