+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -22,6 +22,11 @@
         android:name="com.example.warriorsocial.ui.discover.DiscoverFragment"
         android:label="Discover"
         tools:layout="@layout/act_main" />
+    <fragment
+        android:id="@+id/navigation_notifications"
+        android:name="com.example.warriorsocial.ui.notifications.NotificationsFragment"
+        android:label="Settings"
+        tools:layout="@layout/fragment_settings" />
 
     <fragment
         android:id="@+id/navigation_settings"
